fastpi/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pymupdf==1.24.10
python-pptx==1.0.2
Pillow==10.4.0
requests==2.32.3
llama-index-core==0.10.58
llama-index-readers-file==0.1.30
llama-index-llms-nvidia==0.1.4
llama-index-embeddings-nvidia==0.1.4
llama-index-vector-stores-milvus==0.1.20
pymilvus==2.4.4
python-multipart==0.0.16
boto3
python-multipart
python-dotenv
pydantic
llama-index
transformers
torch
Pillow
python-pptx
pdf2image
pytesseract
pymilvus
snowflake-connector-python
pandas
